+++ imports/api/series/index.js
@@ -30,7 +30,7 @@ if (Meteor.isServer) {
     },
     'series.update' (changes, _id) {
       Series.update({_id: _id}, {$set: changes} );
-      Meteor.call('history.insert', changes, 'series.update');
+      Meteor.call('history.insert', {_id, changes}, 'series.update');
     },
     'series.hide' (_id) {
       var data = { visible: false };