+++ imports/components/Input/File/index.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Meteor } from 'meteor/meteor';
 import tools from '/imports/startup/tools/index';
 
 import Icon from '/imports/components/Icon/index';
@@ -16,6 +15,16 @@ export default class File extends React.Component {
   onChange = (e) => {
     if (e) {
       var files = Array.from(e.target.files);
+      if (files.length > this.props.max) {
+        return alert(`Favor selecionar até ${this.props.max} imagens.`);
+      }
+      if (this.props.allowedFileTypes) {
+        for (var file of files) {
+          if (!this.props.allowedFileTypes.includes(file.type)) {
+            return alert("Favor selecionar um arquivo válido.");
+          }
+        }
+      }
       this.props.onChange(files);
     }
   }
@@ -36,25 +45,58 @@ export default class File extends React.Component {
       )
     }
   }
+  renderPreview = () => {
+    if (!this.props.preview) return null;
+    if (!this.props.value) return null;
+    if (!this.props.value.length) return null;
+    return (
+      <table className="table">
+        <thead>
+          <tr>
+            <th className="table__small-column">#</th>
+            <th>Arquivos Selecionados</th>
+          </tr>
+        </thead>
+        <tbody>
+          {this.props.value.map((file, i) => {
+            return (
+              <tr key={i}>
+                <td>{i+1}</td>
+                <td>{file.name}</td>
+                <td className="table__small-column">
+                  <button onClick={() => this.props.removeFile(i)}>
+                    <Icon icon="not"/>
+                  </button>
+                </td>
+              </tr>
+            )
+          })}
+        </tbody>
+      </table>
+    )
+  }
   render() {
     return (
-      <label className="input__file">
-        <input
-          type="file"
-          files={this.props.files}
-          multiple={this.multiple()}
-          accept={this.props.accept}
-          onChange={this.onChange}
+      <>
+        <label className="input__file">
+          <input
+            type="file"
+            files={this.props.value}
+            multiple={this.multiple()}
+            accept={this.props.accept}
+            onChange={this.onChange}
 
-          readOnly={this.props.readOnly}
-          placeholder={this.props.placeholder}
-          disabled={this.props.disabled}
+            readOnly={this.props.readOnly}
+            placeholder={this.props.placeholder}
+            disabled={this.props.disabled}
 
-          style={this.props.style}
-          />
-        <Icon icon="upload" className="input__file__icon"/>
-        {this.renderText()}
-      </label>
+            style={this.props.style}
+            />
+          <Icon icon="upload" className="input__file__icon"/>
+          {this.renderText()}
+        </label>
+        {this.renderPreview()}
+      </>
     )
   }
 }